requirements.txt
Flask
Flask-WTF
Flask-SQLAlchemy
Flask-Login
Werkzeug
gunicorn
email_validator